+++ modules/core/core_oper/cmd_kill.cpp
@@ -151,7 +151,7 @@ RouteDescriptor CommandKill::GetRouting(User* user, const Params& parameters)
 	return ROUTE_BROADCAST;
 }
 
-void CommandKill::EncodeParameter(std::string& param, unsigned int index)
+void CommandKill::EncodeParameter(std::string& param, size_t index)
 {
 	// Manually translate the nick -> uuid (see above), and also the reason (params[1])
 	// because we decorate it if the oper is local and want remote servers to see the
